containers: remove friendship with template specialization in binary stream class
The friendship is not needed because when a template is declared
friend, every specialization of the template becomes a friend,
whether it is implicitly instantiated, partially specialized,
or explicitly specialized.

Moreover, with the original code, gcc was always using the generic
template implementation of the stream operator although a full
specialization exists for the strings. The same code was working
as expected with clangs.