containers/piercedVector: add 'emreclaim*' functions
These functions will insert a new element in the container. If
the element can reuse an existing position that position will be
initialize using args as the argument for its initialization
otherwise a new element will be created in-place using args as
the arguments for its construction.

These functions are only enabled if the object stored in the
container has an initialization function with a signature like
"void initialize(Args...)".
*
*