containers: rework implementation of the 'PiercedVector' container
The container is split in two components: the kernel and the storage.
The kernel contains the information on the ids stored in the container,
whereas the storage contains the actual data. The PiercedVector is
composed by one kernel and one storage. However, external storage
can be linked with the PiercedVector and kept synchronized with it.